+++ docs/configuration/examples.md
@@ -761,7 +761,7 @@ windows:
 
 tmuxp sessions can be scripted in python. The first way is to use the
 ORM in the {ref}`API`. The second is to pass a {py:obj}`dict` into
-{class}`~tmuxp.workspacebuilder.WorkspaceBuilder` with a correct schema.
+{class}`~tmuxp.workspace.builder.WorkspaceBuilder` with a correct schema.
 See: {meth}`tmuxp.config.validate_schema`.
 
 :::
