+++ tests/test_cli.py
@@ -31,7 +31,7 @@ from tmuxp.cli.utils import (
 )
 from tmuxp.config_reader import ConfigReader
 from tmuxp.workspace import config
-from tmuxp.workspacebuilder import WorkspaceBuilder
+from tmuxp.workspace.builder import WorkspaceBuilder
 
 from .constants import FIXTURE_PATH
 from .fixtures import utils as test_utils
